requirements.txt
pandas>=1.5.0
numpy>=1.22.0
pyarrow>=8.0.0
plotly>=5.0.0
kaleido>=0.2.1
pyyaml>=6.0
openpyxl>=3.0.0
